[WIN32SS][SDK][WIN32NT_APITEST] Move INPUTCONTEXTDX to ntuser.h
