+++ modules/rostests/apitests/win32nt/win32nt.h
@@ -14,6 +14,7 @@
 #include <wingdi.h>
 #include <objbase.h>
 #include <imm.h>
+#include <immdev.h>
 
 #include <winddi.h>
 #include <prntfont.h>
