+++ win32ss/include/ntuser.h
@@ -1244,6 +1244,41 @@ typedef struct tagTRANSMSGLIST
     TRANSMSG TransMsg[ANYSIZE_ARRAY];
 } TRANSMSGLIST, *PTRANSMSGLIST, *LPTRANSMSGLIST;
 
+struct IME_STATE;
+
+#ifdef __cplusplus
+typedef struct INPUTCONTEXTDX : INPUTCONTEXT
+{
+#else
+typedef struct INPUTCONTEXTDX
+{
+    INPUTCONTEXT;
+#endif
+    UINT nVKey;
+    BOOL bNeedsTrans;
+    DWORD dwUnknown1;
+    DWORD dwUIFlags;
+    DWORD dwUnknown2;
+    struct IME_STATE *pState;
+    DWORD dwChange;
+    DWORD dwUnknown5;
+} INPUTCONTEXTDX, *PINPUTCONTEXTDX, *LPINPUTCONTEXTDX;
+
+#ifndef _WIN64
+C_ASSERT(offsetof(INPUTCONTEXTDX, nVKey) == 0x140);
+C_ASSERT(offsetof(INPUTCONTEXTDX, bNeedsTrans) == 0x144);
+C_ASSERT(offsetof(INPUTCONTEXTDX, dwUIFlags) == 0x14c);
+C_ASSERT(offsetof(INPUTCONTEXTDX, pState) == 0x154);
+C_ASSERT(offsetof(INPUTCONTEXTDX, dwChange) == 0x158);
+C_ASSERT(sizeof(INPUTCONTEXTDX) == 0x160);
+#endif
+
+// bits for INPUTCONTEXTDX.dwChange
+#define INPUTCONTEXTDX_CHANGE_OPEN          0x1
+#define INPUTCONTEXTDX_CHANGE_CONVERSION    0x2
+#define INPUTCONTEXTDX_CHANGE_SENTENCE      0x4
+#define INPUTCONTEXTDX_CHANGE_FORCE_OPEN    0x100
+
 #define DEFINE_IME_ENTRY(type, name, params, extended) typedef type (WINAPI *FN_##name) params;
 #include "imetable.h"
 #undef DEFINE_IME_ENTRY
